fuse: add pollhack that compiles on Darwin too
